Add keypad and input handling
